requirements.txt
django-pure-pagination

django >= 2.2
mysqlclient
pillow
django-grappelli
django-haystack
whoosh
jieba